requirements.txt
python-lorem==1.1.2
